requirements.txt
audio-extract
docx2pdf
img2pdf>=0.5.1
opencv-python
openpyxl
pandas
pdf2docx
pdf2image
pikepdf
Pillow
pillow-heif
pytesseract
PyPDF2
pymupdf
rembg
setuptools
tkcolorpicker